IMS_API_Server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django==5.0.4
django-anymail==10.2
django-cors-headers==4.3.1
django-picklefield==3.1
django-q2==1.6.1
django-webpack-loader==3.0.1
djangorestframework==3.14.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itypes==1.2.0
jinja2==3.1.3
markdown==3.5.2
markupsafe==2.1.5
packaging==23.2
postgres==4.0
mysqlclient==2.2.4
psycopg2-binary==2.9.9
psycopg2-pool==1.2
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
setuptools==69.1.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
astroid==3.0.3
autopep8==2.0.4
dill==0.3.8
isort==5.13.2
mccabe==0.7.0
pip-licenses==4.3.4
platformdirs==4.2.0
prettytable==3.10.0
pycodestyle==2.11.1
pylint==3.0.3
tomli==2.0.1
tomlkit==0.12.3
typing-extensions==4.9.0
wcwidth==0.2.13

